+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 - initial release
 
 *0.0.2:*
-- add humans.txt (with bem block for it)
+- merge with project-stub
+- custom level.js for bundles
+- more stable
+
+*0.0.3:*
+- add watcher for simple development
+- add csscomb tool